Initial commit
Initial commit without any sound or timing whatsoever.

Each team is assigned to a key, currently at random.
Game events are translated into a saxophone line on either a jazz minor scale, or a bebop major scale. Range: C#3 - C5
The current pitcher determines the chords being played on the piano - it cycles through the letters in their name, with each letter in the alphabet having a chord assigned to them. Range: C2-C6
The number of baserunners determines the rhythmic pattern of the upright bass, playing the same chord as the piano. Range: C2-C6
TODO: implement sound, duh. But I implemented the music theory successfully!
implement timing within game events
make it into an applet/window with dropdown menus to select the game instead of having to modify the source code to select the game id LMAO

Also included: A hand-rendered mockup of what it should sound like, and a PDF of the score for that mockup, for the bottom of this game: https://reblase.sibr.dev/game/5efd94fd-875f-423e-a47f-cc464d5684e6